update table schema and tag value

+++ src/inc/taosmsg.h
@@ -284,7 +284,7 @@ typedef struct {
   int32_t   tid;
   int16_t   tversion;
   int16_t   colId;
-  int16_t   type;
+  int8_t    type;
   int16_t   bytes;
   int32_t   tagValLen;
   int16_t   numOfTags;
